Updated SearchArtist UseCase, segregated UseCases
