+++ app/src/main/java/com/chnouman/lastfmapidemo/di/UseCasesModule.kt
@@ -3,13 +3,13 @@ package com.chnouman.lastfmapidemo.di
 import com.chnouman.lastfmapidemo.domain.repository.MainRepository
 import com.chnouman.lastfmapidemo.domain.repository.SearchRepository
 import com.chnouman.lastfmapidemo.domain.repository.TopAlbumRepository
-import com.chnouman.lastfmapidemo.domain.usecases.AddAlbum
-import com.chnouman.lastfmapidemo.domain.usecases.CompareLocalAlbums
-import com.chnouman.lastfmapidemo.domain.usecases.DeleteAlbum
-import com.chnouman.lastfmapidemo.domain.usecases.GetAlbumInfo
-import com.chnouman.lastfmapidemo.domain.usecases.GetLocalAlbums
-import com.chnouman.lastfmapidemo.domain.usecases.GetTopAlbums
-import com.chnouman.lastfmapidemo.domain.usecases.SearchArtist
+import com.chnouman.lastfmapidemo.domain.usecases.album.AddAlbum
+import com.chnouman.lastfmapidemo.domain.usecases.album.CompareLocalAlbums
+import com.chnouman.lastfmapidemo.domain.usecases.album.DeleteAlbum
+import com.chnouman.lastfmapidemo.domain.usecases.album.GetAlbumInfo
+import com.chnouman.lastfmapidemo.domain.usecases.album.GetLocalAlbums
+import com.chnouman.lastfmapidemo.domain.usecases.album.GetTopAlbums
+import com.chnouman.lastfmapidemo.domain.usecases.artist.SearchArtist
 import com.chnouman.lastfmapidemo.domain.usecases.artist.AddArtist
 import com.chnouman.lastfmapidemo.domain.usecases.artist.DeleteArtist
 import com.chnouman.lastfmapidemo.domain.usecases.track.AddTrack
