--- app/src/main/java/com/chnouman/lastfmapidemo/domain/usecases/SearchArtist.kt
@@ -1,19 +1,0 @@
-package com.chnouman.lastfmapidemo.domain.usecases
-
-import com.chnouman.lastfmapidemo.BuildConfig
-import com.chnouman.lastfmapidemo.core.util.Resource
-import com.chnouman.lastfmapidemo.data.local.entities.Artist
-import com.chnouman.lastfmapidemo.domain.repository.SearchRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-
-class SearchArtist(private val repository: SearchRepository) {
-
-    suspend operator fun invoke(word: String): Flow<Resource<List<Artist>>> {
-        return if (word.isEmpty()) {
-            flow { }
-        } else {
-            return repository.searchArtist(word, BuildConfig.API_KEY)
-        }
-    }
-}
