+++ app/src/main/java/com/chnouman/lastfmapidemo/domain/usecases/album/DeleteAlbum.kt
@@ -1,10 +1,10 @@
-package com.chnouman.lastfmapidemo.domain.usecases
+package com.chnouman.lastfmapidemo.domain.usecases.album
 
 import com.chnouman.lastfmapidemo.data.local.entities.Album
 import com.chnouman.lastfmapidemo.domain.repository.TopAlbumRepository
 
-class AddAlbum(private val repository: TopAlbumRepository) {
+class DeleteAlbum(private val repository: TopAlbumRepository) {
     suspend operator fun invoke(albumsDto: Album) {
-        return repository.addAlbumDto(albumsDto)
+        return repository.deleteAlbum(albumsDto)
     }
 }
